chg ! use chunks

+++ src/hope_dedup_engine/constants.py
@@ -2,10 +2,10 @@ from enum import Enum
 
 
 class FacialError(Enum):
-    GENERIC_ERROR = 999.0
-    NO_FACE_DETECTED = 998.0
-    MULTIPLE_FACES_DETECTED = 997.0
-    NO_FILE_FOUND = 996.0
+    GENERIC_ERROR = 999
+    NO_FACE_DETECTED = 998
+    MULTIPLE_FACES_DETECTED = 997
+    NO_FILE_FOUND = 996
 
     @property
     def code(self) -> int:
@@ -15,14 +15,6 @@ class FacialError(Enum):
 def is_facial_error(value):
     if isinstance(value, str):
         return value in FacialError.__members__
-    if isinstance(value, float):
+    if isinstance(value, int):
         return value in FacialError._value2member_map_
     return False
-
-
-# NO_FACE_DETECTED: Final[str] = "NO_FACE_DETECTED"
-# MULTIPLE_FACES_DETECTED: Final[str] = "MULTIPLE_FACE_DETECTED"
-# FILE_ERROR: Final[str] = "GENERIC_ERROR"
-# ERRORS: Final[list[str]] = [NO_FACE_DETECTED, MULTIPLE_FACES_DETECTED, FILE_ERROR]
-
-# NO_ENCODING: Final[float] = 999
